+++ database/migrations/2020_10_12_161747_create_attendances_table.php
@@ -24,8 +24,8 @@ class CreateAttendancesTable extends Migration
             $table->time('in_time_2')->nullable();
             $table->time('out_time_1')->nullable();
             $table->time('out_time_2')->nullable();
-            $table->float('ot')->nullable();
-            $table->float('ot_extra')->nullable();
+            $table->double('ot')->nullable();
+            $table->double('ot_extra')->nullable();
 
             $table->unsignedBigInteger('user_id')->nullable();
             $table->timestamps();
